issue#1029: replicate rights to all references
When doing the "manual" rights distribution
(i.e. not for categories, but single contents and object blocks)
the replication did transfer the rights not to all
references of the given object, but only the last one.
 The current version transfers the right to all references,
even though it uses only one source (always the given ref).
Different right settings for different references are not supported.

Also add a big "if (true)" around the other role sync stuff
which avoids syncing the other righs stuff if set to false
(and the contents to be synced are given "manually").